Known error serialization with nested cause (#56)
Implements the changes mentioned in
https://github.com/nexus-rpc/api/pull/21.
I've added tests in the last commit that should not be merged, they are
used to test current main server and client against this version to
ensure compatibility.
Note that backwards compatibility is only ensured if an error includes
only `Cause` or `Message` but not both, but that is okay as the new
`Message` field is an addition.

I will wait on merging this until the transport refactor is in to avoid
merge conflicts in that branch.