Remove client and server implementations, and OperationHandler GetInfo, and GetResult (#62)
Also deprecate `Handler.GetInfo` and `Handler.GetResult`, those are in
use by the Temporal Go SDK and cannot be removed for now.
Also revert the error message changes, the Temporal server needs to be
adjusted to support that still.
The client and server implementations will be inlined into
`temporalio/temporal` until the protocol matures.